Create update permissions backend logic

+++ backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -16,7 +17,7 @@ const Mutations = {
           // create relationship between the item and the user
           user: {
             connect: {
-              id: ctx.request.userIda,
+              id: ctx.request.userId,
             },
           },
           ...args,
@@ -46,9 +47,15 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // find the item
-    const item = await ctx.db.query.item({ where }, `{ id, title }`);
-    // check if they have permissions
-    // TODO
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+    // TODO: check if they have permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to delete this item");
+    }
     // delete item
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -166,6 +173,35 @@ const Mutations = {
     // return the new user
 
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // check it user is logged in
+    if (!ctx.request.userId) {
+      throw new Error(`You must be logged in!`);
+    }
+    // query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId,
+        },
+      },
+      info
+    );
+    // check if they have the pernussuibs to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: { id: args.userId },
+      },
+      info
+    );
   },
 };
